UISystem overhaul
Added Textfield.
Added InputModes which are used by the UISystem to intercept input
for the game world or for the UI.
Added UIInputConfig which is used by UISystem to filter which
elements receive what types of events.
Added `Base.show` functions for UIElements.
Fixed an incorrect method call in Labels.